+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Export logs to Jaeger ([#360]).
 - Added common datastructures shared between all operators like `Tls` oder `AuthenticationClass` ([#366]).
+- Added helpers for env variables from Secrets or ConfigMaps ([#370]).
 
 ### Changed
 
@@ -17,6 +18,7 @@ All notable changes to this project will be documented in this file.
 [#360]: https://github.com/stackabletech/operator-rs/pull/360
 [#366]: https://github.com/stackabletech/operator-rs/pull/366
 [#369]: https://github.com/stackabletech/operator-rs/pull/369
+[#370]: https://github.com/stackabletech/operator-rs/pull/370
 
 ## [0.15.0] - 2022.03.21